pyproject.toml
[project]
name = "optics-raytracer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "KoStard", email = "ruben.kostard@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
]

[project.scripts]
optics-raytracer = "optics_raytracer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
